# openapi.yaml
openapi: 3.0.3
info:
  title: Spacecraft API
  version: 1.0.0
  description: Planogram management system API
paths:
...
  /api/planograms/{slug}/ai-overview/:
    post:
      operationId: planograms_ai_overview_create
      description: Generate AI-powered overview of the planogram layout.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - planograms
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIOverviewResponse'
          description: ''
  /api/planograms/{slug}/layout/:
    get:
      operationId: planograms_layout_retrieve
      description: Get planogram layout.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - planograms
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
          description: ''
    post:
      operationId: planograms_layout_create
      description: Save planogram layout and return updated planogram data.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - planograms
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanogramLayoutRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Planogram'
          description: ''
  /api/planograms/{slug}/layout/add-products/:
    post:
      operationId: planograms_layout_add_products_create
      description: Add products to planogram layout at specified row.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - planograms
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProductsRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
          description: ''
...
components:
  schemas:
    AddProductItem:
      type: object
      description: Serializer for individual product item in add products request.
      properties:
        id:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
      required:
      - id
      - quantity
    AddProductsRequest:
      type: object
      description: Serializer for add products to layout request.
      properties:
        row_id:
          type: integer
          minimum: 0
        products:
          type: array
          items:
            $ref: '#/components/schemas/AddProductItem'
      required:
      - products
      - row_id
    AIOverviewResponse:
      type: object
      description: Output serializer for AI overview response.
      properties:
        overview:
          type: string
      required:
      - overview
...
    GridConfig:
      type: object
      description: Serializer for grid configuration.
      properties:
        cols:
          type: integer
        rows:
          type: integer
        cellWidthIn:
          type: number
          format: double
        normalizedWidthIn:
          type: number
          format: double
        normalizedHeightIn:
          type: number
          format: double
      required:
      - cellWidthIn
      - cols
      - normalizedHeightIn
      - normalizedWidthIn
      - rows
    GridConfigRequest:
      type: object
      description: Serializer for grid configuration.
      properties:
        cols:
          type: integer
        rows:
          type: integer
        cellWidthIn:
          type: number
          format: double
        normalizedWidthIn:
          type: number
          format: double
        normalizedHeightIn:
          type: number
          format: double
      required:
      - cellWidthIn
      - cols
      - normalizedHeightIn
      - normalizedWidthIn
      - rows
    Layout:
      type: object
      description: Serializer for complete planogram layout structure.
      properties:
        grid:
          $ref: '#/components/schemas/GridConfig'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/LayoutRow'
      required:
      - grid
      - rows
    LayoutItem:
      type: object
      description: Serializer for individual layout items (grid items).
      properties:
        i:
          type: string
        x:
          type: integer
        y:
          type: integer
        w:
          type: integer
        h:
          type: integer
        meta:
          $ref: '#/components/schemas/LayoutItemMeta'
      required:
      - h
      - i
      - meta
      - w
      - x
      - y
    LayoutItemMeta:
      type: object
      description: Serializer for layout item metadata.
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        color:
          type: string
        score:
          type: number
          format: double
        pack_width_in:
          type: number
          format: double
        pack_height_in:
          type: number
          format: double
      required:
      - category
      - id
      - name
      - pack_height_in
      - pack_width_in
      - score
    LayoutItemMetaRequest:
      type: object
      description: Serializer for layout item metadata.
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        color:
          type: string
        score:
          type: number
          format: double
        pack_width_in:
          type: number
          format: double
        pack_height_in:
          type: number
          format: double
      required:
      - category
      - id
      - name
      - pack_height_in
      - pack_width_in
      - score
    LayoutItemRequest:
      type: object
      description: Serializer for individual layout items (grid items).
      properties:
        i:
          type: string
          minLength: 1
        x:
          type: integer
        y:
          type: integer
        w:
          type: integer
        h:
          type: integer
        meta:
          $ref: '#/components/schemas/LayoutItemMetaRequest'
      required:
      - h
      - i
      - meta
      - w
      - x
      - y
    LayoutRequest:
      type: object
      description: Serializer for complete planogram layout structure.
      properties:
        grid:
          $ref: '#/components/schemas/GridConfigRequest'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/LayoutRowRequest'
      required:
      - grid
      - rows
    LayoutRow:
      type: object
      description: Serializer for layout rows (shelves).
      properties:
        id:
          type: integer
        category:
          type: string
          nullable: true
        name:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/LayoutItem'
      required:
      - category
      - id
      - items
      - name
    LayoutRowRequest:
      type: object
      description: Serializer for layout rows (shelves).
      properties:
        id:
          type: integer
        category:
          type: string
          nullable: true
          minLength: 1
        name:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/LayoutItemRequest'
      required:
      - category
      - id
      - items
      - name
...
    Planogram:
      type: object
      description: Output serializer for Planogram model.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          readOnly: true
        season:
          $ref: '#/components/schemas/SeasonC9aEnum'
        project:
          type: integer
        project_name:
          type: string
          readOnly: true
        project_slug:
          type: string
          readOnly: true
        display:
          type: integer
          description: Reference to display used for dimensions (auto-selected on
            creation)
        display_name:
          type: string
          readOnly: true
        company:
          type: integer
          readOnly: true
        company_name:
          type: string
          readOnly: true
        width_in:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          description: Display width in inches
        height_in:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          description: Display height in inches
        depth_in:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          description: Display depth in inches
        shelf_count:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          description: Number of shelves
        shelf_spacing:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          description: Shelf spacing in inches
        category_ids:
          type: array
          items:
            type: integer
          readOnly: true
        categories:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: Derive category names from category_ids.
          readOnly: true
        created_by:
          type: integer
          readOnly: true
        created_by_username:
          type: string
          readOnly: true
        updated_by:
          type: integer
          readOnly: true
        updated_by_username:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - categories
      - category_ids
      - company
      - company_name
      - created_at
      - created_by
      - created_by_username
      - depth_in
      - display
      - display_name
      - height_in
      - id
      - name
      - project
      - project_name
      - project_slug
      - shelf_count
      - shelf_spacing
      - slug
      - updated_at
      - updated_by
      - updated_by_username
      - width_in
...
    PlanogramLayoutRequest:
      type: object
      description: Input serializer for layout save.
      properties:
        layout:
          $ref: '#/components/schemas/LayoutRequest'
      required:
      - layout
...
    SeasonC9aEnum:
      enum:
      - spring
      - summer
      - fall
      - winter
      type: string
      description: |-
        * `spring` - Spring
        * `summer` - Summer
        * `fall` - Fall
        * `winter` - Winter
...